+++ CHANGELOG.md
@@ -31,6 +31,8 @@ changes. Best viewed [here](https://google-grain.readthedocs.io/en/latest/change
     specification inference.
   * Adds support for changing `IterDataset.mix` components and weights after a
     checkpoint.
+  * Adds experimental support for `get_next_index` and `set_next_index` to fetch
+    and advance a `grain.DatasetIterator` to the given produced element index.
 
 ## Grain 0.2.14 (October 30, 2025)